% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dDAGreverse}
\alias{dDAGreverse}
\title{Function to reverse the edge direction of a direct acyclic graph (DAG)}
\usage{
dDAGreverse(g)
}
\arguments{
\item{g}{an object of class "igraph" or "graphNEL"}
}
\value{
\itemize{
\item{\code{gr}: a graph being reversed, an object of class "igraph" or
"graphNEL"}
}
}
\description{
\code{dDAGreverse} is supposed to reverse the edge direction of a
direct acyclic graph (DAG; an ontology). The return graph remains all
attributes associated on nodes and edges.
}
\note{
none
}
\examples{
# 1) load HPPA as igraph object
data(ig.HPPA)
g <- ig.HPPA

# 2) the graph with reverse edge direction
gr <- dDAGreverse(g)
gr
}
\seealso{
\code{\link{dDAGreverse}}
}

